# Conflicts:
#	src/main/java/de/uniks/beastopia/teaml/controller/ingame/encounter/EncounterController.java
#	src/main/java/de/uniks/beastopia/teaml/controller/ingame/encounter/RenderBeastController.java

+++ src/main/java/de/uniks/beastopia/teaml/rest/Trainer.java
@@ -13,7 +13,7 @@ public record Trainer(
         String image,
         List<String> team,
         List<String> visitedAreas,
-        int coins,
+        float coins,
         String area,
         int x,
         int y,
